Refactor split events/commands
