+++ app/Auth0/Jobs/BlockClientListener.php
@@ -4,13 +4,16 @@ declare(strict_types=1);
 
 namespace App\Auth0\Jobs;
 
+use App\Auth0\Auth0ClusterSDK;
 use App\Auth0\Events\BlockClient;
 use App\Auth0\Events\ClientBlocked;
+use App\Auth0\Repositories\Auth0ClientRepository;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
+use Psr\Log\LoggerInterface;
 
 final class BlockClientListener implements ShouldQueue
 {
@@ -19,8 +22,31 @@ final class BlockClientListener implements ShouldQueue
     use Queueable;
     use SerializesModels;
 
-    public function handle(BlockClient $blockClient): void
+    public function __construct(
+        private readonly Auth0ClusterSDK $clusterSDK,
+        private readonly Auth0ClientRepository $auth0ClientRepository,
+        private readonly LoggerInterface $logger,
+    ) {
+    }
+
+    public function handle(BlockClient $event): void
     {
-        ClientBlocked::dispatch($blockClient->id);
+        $oauth0Client = $this->auth0ClientRepository->getById($event->id);
+
+        if ($oauth0Client === null) {
+            return;
+        }
+
+        $this->clusterSDK->blockClients($oauth0Client);
+
+        $this->logger->info(
+            'Auth0 client blocked',
+            [
+                'domain' => 'auth0',
+                'id' => $event->id,
+            ]
+        );
+
+        ClientBlocked::dispatch($event->id);
     }
 }
