+++ app/Nova/ActionGuards/ActionGuard.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\Nova\ActionGuards;
 
-interface ActionGuardInterface
+interface ActionGuard
 {
     public function canDo(object $resource): bool;
 }
